Correct buildability after moving some mock-classes
