+++ src/components/hmi_message_handler/test/include/hmi_message_handler/mock_hmi_message_observer.h
@@ -39,19 +39,22 @@
 #include "hmi_message_handler/hmi_message_observer.h"
 #include "application_manager/message.h"
 
-namespace hmi_message_handler {
+namespace test {
+namespace components {
+namespace hmi_message_handler_test {
 
 using ::hmi_message_handler::HMIMessageObserver;
 
-class MockHMIMessageObserver : public HMIMessageObserver,
-                               public utils::Singleton<MockHMIMessageObserver> {
+class MockHMIMessageObserver : public HMIMessageObserver {
  public:
   MOCK_METHOD1(OnMessageReceived,
                void(utils::SharedPtr<application_manager::Message> message));
   MOCK_METHOD1(OnErrorSending,
                void(utils::SharedPtr<application_manager::Message> message));
-  virtual ~MockHMIMessageObserver() {}
 };
+
 }  // namespace hmi_message_handler
+}  // namespace components
+}  // namespace test
 
 #endif  // SRC_COMPONENTS_HMI_MESSAGE_HANDLER_TEST_INCLUDE_HMI_MESSAGE_HANDLER_MOCK_HMI_MESSAGE_OBSERVER_H_
