+++ src/components/transport_manager/test/include/transport_manager/mock_telemetry_observer.h
@@ -40,7 +40,7 @@ namespace test {
 namespace components {
 namespace transport_manager_test {
 
-class MockTMTelemetryObserver : public TMTelemetryObserver {
+class MockTMTelemetryObserver : public transport_manager::TMTelemetryObserver {
  public:
   MOCK_METHOD1(StartRawMsg,
       void(const protocol_handler::RawMessage* ptr));
